+++ src/page/etc/index.js
@@ -1,0 +1,1 @@
+export { default as NotFoundPage } from './NotFoundPage/NotFoundPage.jsx';
