--- src/pages/NotFoundPage/index.js
@@ -1,1 +1,0 @@
-export { default as NotFoundPage } from './NotFoundPage.jsx';
